Interactive testing file.
File containing testing for Oscillioscope and Spectrum Analysis on a given wave file. In this Python notebook the system audio of the machine was used through the use of Virtual Audio Cable. The testing will go on to test in real time and a discussion will be made about switching matplotlib to another graphing module.