Added new graph.
New graph of Frequency x Amplitude added. A result of this is a huge increase in CPU usage. This needs to be optimised. I have a feeling that a fair amount of CPU usage could be reduced if the second graph can be updated in the same way as the others, instead of just updating the line.